requirements.txt
cssselect==1.2.0
decorator==5.1.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.2.10
django-filter==24.2
django-import-export==4.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==1.0.0
fake-useragent==1.1.0
filelock==3.13.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0

importlib-metadata==5.1.0
importlib-resources==5.10.0
ipython==8.5.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
numpy==2.0.0
pandas==2.2.2
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3

pyee==8.2.2
Pygments==2.13.0

pyppeteer==1.0.2
pyquery==1.4.3

python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1

requests-html==0.10.0


sniffio==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.5.0
tablib==3.5.0

traitlets==5.4.0
typing_extensions==4.10.0
tzdata==2024.1

virtualenv==20.25.1
w3lib==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
zipp==3.11.0
